Initial separation of Navigator -> ActivityReport
