+++ frontend/src/components/MultiSelect.js
@@ -24,7 +24,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import Select, { components } from 'react-select';
 import Creatable from 'react-select/creatable';
-import { Controller } from 'react-hook-form/dist/index.ie11';
+import { Controller } from 'react-hook-form';
 import _ from 'lodash';
 import { v4 as uuidv4 } from 'uuid';
 import arrowBoth from '../images/arrow-both.svg';
