+++ frontend/src/pages/ActivityReport/Pages/Review/Approver/Review.js
@@ -1,7 +1,7 @@
 import React, { useContext } from 'react';
 import moment from 'moment';
 import PropTypes from 'prop-types';
-import { useFormContext } from 'react-hook-form/dist/index.ie11';
+import { useFormContext } from 'react-hook-form';
 import _ from 'lodash';
 import {
   Dropdown, Form, Label, Fieldset, Button,
