+++ frontend/src/pages/ActivityReport/Pages/Review/ReviewPage.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { some } from 'lodash';
-import { useFormContext } from 'react-hook-form/dist/index.ie11';
+import { useFormContext } from 'react-hook-form';
 
 import Section from './ReviewSection';
 import ReviewItem from './ReviewItem';
