+++ frontend/src/pages/ActivityReport/Pages/Review/Submitter/__tests__/index.js
@@ -6,7 +6,7 @@ import userEvent from '@testing-library/user-event';
 import React from 'react';
 import { Router } from 'react-router';
 import { createMemoryHistory } from 'history';
-import { useForm, FormProvider } from 'react-hook-form/dist/index.ie11';
+import { useForm, FormProvider } from 'react-hook-form';
 import Submitter from '../index';
 import NetworkContext from '../../../../../../NetworkContext';
