+++ frontend/src/pages/ActivityReport/Pages/__tests__/nextSteps.js
@@ -5,7 +5,7 @@ import {
   render, screen, waitFor, fireEvent,
 } from '@testing-library/react';
 import React from 'react';
-import { FormProvider, useForm } from 'react-hook-form/dist/index.ie11';
+import { FormProvider, useForm } from 'react-hook-form';
 
 import nextSteps, { isPageComplete } from '../nextSteps';
