+++ frontend/src/pages/ActivityReport/Pages/__tests__/supportingAttachments.js
@@ -2,7 +2,7 @@
 import '@testing-library/jest-dom';
 import { render, screen } from '@testing-library/react';
 import React from 'react';
-import { FormProvider, useForm } from 'react-hook-form/dist/index.ie11';
+import { FormProvider, useForm } from 'react-hook-form';
 import { Router } from 'react-router-dom';
 import { createMemoryHistory } from 'history';
