+++ frontend/src/pages/ActivityReport/Pages/activitySummary.js
@@ -3,7 +3,7 @@ import React, {
 } from 'react';
 import PropTypes from 'prop-types';
 import { Helmet } from 'react-helmet';
-import { useFormContext } from 'react-hook-form/dist/index.ie11';
+import { useFormContext } from 'react-hook-form';
 import { isEmpty, isUndefined } from 'lodash';
 import {
   Fieldset, Radio, Grid, TextInput, Checkbox, Label,
@@ -29,6 +29,7 @@ import HtmlReviewItem from './Review/HtmlReviewItem';
 import Section from './Review/ReviewSection';
 import { reportIsEditable } from '../../../utils';
 import IndicatesRequiredField from '../../../components/IndicatesRequiredField';
+import NavigatorButtons from '../../../components/Navigator/components/NavigatorButtons';
 
 const ActivitySummary = ({
   recipients,
@@ -644,13 +645,31 @@ export default {
   path: 'activity-summary',
   reviewSection: () => <ReviewSection />,
   review: false,
-  render: (additionalData) => {
+  render: (
+    additionalData,
+    _formData,
+    _reportId,
+    isAppLoading,
+    onContinue,
+    onSaveDraft,
+    onUpdatePage,
+  ) => {
     const { recipients, collaborators } = additionalData;
     return (
-      <ActivitySummary
-        recipients={recipients}
-        collaborators={collaborators}
-      />
+      <>
+        <ActivitySummary
+          recipients={recipients}
+          collaborators={collaborators}
+        />
+        <NavigatorButtons
+          isAppLoading={isAppLoading}
+          onContinue={onContinue}
+          onSaveDraft={onSaveDraft}
+          path="activity-summary"
+          position={1}
+          onUpdatePage={onUpdatePage}
+        />
+      </>
     );
   },
   isPageComplete,
