+++ frontend/src/pages/ActivityReport/Pages/components/OtherEntity.js
@@ -2,7 +2,7 @@ import React, {
   useState, useEffect,
 } from 'react';
 import PropTypes from 'prop-types';
-import { useFormContext, useFieldArray } from 'react-hook-form/dist/index.ie11';
+import { useFormContext, useFieldArray } from 'react-hook-form';
 import { Alert } from '@trussworks/react-uswds';
 import { DECIMAL_BASE } from '@ttahub/common';
 import Objective from './Objective';
@@ -13,7 +13,7 @@ import { NEW_OBJECTIVE } from './constants';
 const OBJECTIVE_LABEL = 'objectivesWithoutGoals';
 
 export default function OtherEntity({
-  recipientIds, onSaveDraft, reportId,
+  recipientIds, reportId,
 }) {
   const { errors } = useFormContext();
   const [topicOptions, setTopicOptions] = useState([]);
@@ -69,7 +69,6 @@ export default function OtherEntity({
             errors={objectiveErrors}
             remove={remove}
             fieldArrayName={OBJECTIVE_LABEL}
-            onSaveDraft={onSaveDraft}
             reportId={parseInt(reportId, DECIMAL_BASE)}
           />
         );
@@ -81,6 +80,5 @@ export default function OtherEntity({
 
 OtherEntity.propTypes = {
   recipientIds: PropTypes.arrayOf(PropTypes.number).isRequired,
-  onSaveDraft: PropTypes.func.isRequired,
   reportId: PropTypes.string.isRequired,
 };
