+++ frontend/src/pages/ActivityReport/Pages/components/__tests__/ConditionalFieldsForHookForm.js
@@ -7,7 +7,7 @@ import {
 import React from 'react';
 import {
   FormProvider, useForm,
-} from 'react-hook-form/dist/index.ie11';
+} from 'react-hook-form';
 import ConditionalFieldsForHookForm from '../ConditionalFieldsForHookForm';
 
 const DEFAULT_PROMPTS = [{
