+++ frontend/src/pages/ActivityReport/Pages/components/__tests__/ResourceSelector.js
@@ -3,7 +3,7 @@ import '@testing-library/jest-dom';
 import userEvent from '@testing-library/user-event';
 import { render, screen, waitFor } from '@testing-library/react';
 import React from 'react';
-import { FormProvider, useForm } from 'react-hook-form/dist/index.ie11';
+import { FormProvider, useForm } from 'react-hook-form';
 
 import ResourceSelector from '../ResourceSelector';
