+++ frontend/src/pages/ActivityReport/Pages/nextSteps.js
@@ -6,6 +6,7 @@ import { Fieldset } from '@trussworks/react-uswds';
 import NextStepsRepeater from './components/NextStepsRepeater';
 import ReviewPage from './Review/ReviewPage';
 import IndicatesRequiredField from '../../../components/IndicatesRequiredField';
+import NavigatorButtons from '../../../components/Navigator/components/NavigatorButtons';
 
 export const isPageComplete = (formData, formState) => {
   const { isValid } = formState;
@@ -93,9 +94,29 @@ export default {
   path: 'next-steps',
   review: false,
   reviewSection: () => <ReviewSection />,
-  render: (_additionalData, formData) => {
+  render: (
+    _additionalData,
+    formData,
+    _reportId,
+    isAppLoading,
+    onContinue,
+    onSaveDraft,
+    onUpdatePage,
+  ) => {
     const { activityRecipientType } = formData;
-    return (<NextSteps activityRecipientType={activityRecipientType} />);
+    return (
+      <>
+        <NextSteps activityRecipientType={activityRecipientType} />
+        <NavigatorButtons
+          isAppLoading={isAppLoading}
+          onContinue={onContinue}
+          onSaveDraft={onSaveDraft}
+          onUpdatePage={onUpdatePage}
+          path="next-steps"
+          position={4}
+        />
+      </>
+    );
   },
   isPageComplete,
 };
