+++ frontend/src/pages/ActivityReport/Pages/supportingAttachments.js
@@ -1,15 +1,18 @@
 import React, { useState } from 'react';
 import PropTypes from 'prop-types';
-import { Controller, useFormContext } from 'react-hook-form/dist/index.ie11';
+import { Controller, useFormContext } from 'react-hook-form';
 import { Helmet } from 'react-helmet';
 import {
   ErrorMessage,
-  Fieldset, FormGroup, Label,
+  Fieldset,
+  FormGroup,
+  Label,
 } from '@trussworks/react-uswds';
 import Section from './Review/ReviewSection';
 import FileReviewItem from './Review/FileReviewItem';
 import { reportIsEditable } from '../../../utils';
 import ActivityReportFileUploader from '../../../components/FileUploader/ActivityReportFileUploader';
+import NavigatorButtons from '../../../components/Navigator/components/NavigatorButtons';
 
 const SupportingAttachments = ({
   reportId,
@@ -93,9 +96,27 @@ export default {
   path: 'supporting-attachments',
   reviewSection: () => <ReviewSection />,
   review: false,
-  render: (_additionalData, _formData, reportId) => (
-    <SupportingAttachments
-      reportId={reportId}
-    />
+  render: (
+    _additionalData,
+    _formData,
+    reportId,
+    isAppLoading,
+    onContinue,
+    onSaveDraft,
+    onUpdatePage,
+  ) => (
+    <>
+      <SupportingAttachments
+        reportId={reportId}
+      />
+      <NavigatorButtons
+        isAppLoading={isAppLoading}
+        onContinue={onContinue}
+        onSaveDraft={onSaveDraft}
+        onUpdatePage={onUpdatePage}
+        path="supporting-attachments"
+        position={3}
+      />
+    </>
   ),
 };
